Improve QR codes on screen
When provisioning is done, the QR code will not be shown anymore.
This would happen before this change, and be almost immediately overwritten with the info QR code.

The info QR code is shown immediately after we know the device is provisioned.
This is to minimize the delay between clearing the screen (on init) and showing the QR.
This will happen every time the device boots, so now it will be less noticable.