Send a heartbeat once after provisioning
Trello: https://trello.com/c/zU8GZMZe

Previously every measurement was done once and then sent.
When provisioning a new device with the WarmteWachter app,
this would cause a long delay before the final checkmark would appear.

This change makes sure that a heartbeat is sent right after provisioning.
Only after the heartbeat was successufully sent,
will the rest of the measurements be sent.
This makes sure that the final checkmark in the WarmeteWachter app
is quickly checked.