Turn button press handler into non-class
Before this change, a new handler would have to be created for each translation unit.
These handlers and their interrupts would conflict with each other.

This change makes sure there is only one handler, and prevents these issues.

Some dependent code also had to be changed in order to adapt to this.