Enable power off with long press of power button
The device was unable to be turned off manually before this change.

This gave a lot of problems if the screen had been in direct sunlight.
This would cause the QR-code to become unreadable.

This change makes it possible to manually turn the device off.
The device will buzz for 500ms to signal powering off.