Display text under info QR code
Trello: https://trello.com/c/dE7AePCQ

This change required the use of LovyanGFX directly.
We previously used it through twomes_provisioning_tools.
That library did not have support for displaying text.

A new class 'Screen' was created, which can control an eInk screen.
It can clear the screen and display a QR, with or without text.
If the text is changed, the text will resize to fit under a QR code.

The text 'Scan voor info' will now be displayed under the info QR code after provisioning.