Implement true appending of measurements
Previously, all measurements in SPIFFS had to be loaded into RAM,
then new measurements appended and written to SPIFFS.

This commit immediately appends items to the file in SPIFFS
without loading the old items.

This change was made possibly by not saving the amount of measurements,
but continually reading items
and stopping when we reach the end of the file.