feat: exit BT screen after timeout
A timer is reset when a button is pressed in the BT menu.
If no buttons are pressed for a set amount of time, we exit the menu.

The timer is implemented as a FreeRTOS task that delays.
That means the timeout is not accurate down to the millisecond
if other high-priority tasks are running.
It was implemented this way because
using the esp_timer_... functions cause lovyanGFX to crash the ESP.